Break security group rules out into a separate resource
This allows security groups to have circular dependencies on each other.
