+++ src/main/java/gyro/aws/ec2/SecurityGroupResource.java
@@ -16,31 +16,24 @@
 
 package gyro.aws.ec2;
 
+import java.util.Set;
+import java.util.concurrent.TimeUnit;
+
 import com.psddev.dari.util.ObjectUtils;
 import gyro.aws.AwsResource;
 import gyro.aws.Copyable;
 import gyro.core.GyroException;
 import gyro.core.GyroUI;
+import gyro.core.Type;
 import gyro.core.Wait;
 import gyro.core.resource.Id;
-import gyro.core.resource.Updatable;
-import gyro.core.Type;
 import gyro.core.resource.Output;
 import gyro.core.scope.State;
 import software.amazon.awssdk.services.ec2.Ec2Client;
 import software.amazon.awssdk.services.ec2.model.CreateSecurityGroupResponse;
 import software.amazon.awssdk.services.ec2.model.DescribeSecurityGroupsResponse;
 import software.amazon.awssdk.services.ec2.model.Ec2Exception;
-import software.amazon.awssdk.services.ec2.model.IpPermission;
-import software.amazon.awssdk.services.ec2.model.IpRange;
-import software.amazon.awssdk.services.ec2.model.Ipv6Range;
 import software.amazon.awssdk.services.ec2.model.SecurityGroup;
-import software.amazon.awssdk.services.ec2.model.UserIdGroupPair;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
-import java.util.concurrent.TimeUnit;
 
 /**
  * Create a security group with specified rules.
@@ -70,9 +63,6 @@ public class SecurityGroupResource extends Ec2TaggableResource<SecurityGroup> im
     private String name;
     private VpcResource vpc;
     private String description;
-    private List<SecurityGroupIngressRuleResource> ingress;
-    private List<SecurityGroupEgressRuleResource> egress;
-    private Boolean keepDefaultEgressRules;
 
     // Read-only
     private String id;
@@ -112,56 +102,6 @@ public class SecurityGroupResource extends Ec2TaggableResource<SecurityGroup> im
     }
 
     /**
-     * A list of ingress rules to block inbound traffic.
-     *
-     * @subresource gyro.aws.ec2.SecurityGroupIngressRuleResource
-     */
-    public List<SecurityGroupIngressRuleResource> getIngress() {
-        if (ingress == null) {
-            ingress = new ArrayList<>();
-        }
-
-        return ingress;
-    }
-
-    public void setIngress(List<SecurityGroupIngressRuleResource> ingress) {
-        this.ingress = ingress;
-    }
-
-    /**
-     * A list of egress rules to block outbound traffic.
-     *
-     * @subresource gyro.aws.ec2.SecurityGroupEgressRuleResource
-     */
-    public List<SecurityGroupEgressRuleResource> getEgress() {
-        if (egress == null) {
-            egress = new ArrayList<>();
-        }
-
-        return egress;
-    }
-
-    public void setEgress(List<SecurityGroupEgressRuleResource> egress) {
-        this.egress = egress;
-    }
-
-    /**
-     * Whether to keep the default egress rule. If false, the rule will be deleted.
-     */
-    @Updatable
-    public boolean isKeepDefaultEgressRules() {
-        if (keepDefaultEgressRules == null) {
-            keepDefaultEgressRules = true;
-        }
-
-        return keepDefaultEgressRules;
-    }
-
-    public void setKeepDefaultEgressRules(boolean keepDefaultEgressRules) {
-        this.keepDefaultEgressRules = keepDefaultEgressRules;
-    }
-
-    /**
      * The ID of the Security Group.
      */
     @Id
@@ -198,68 +138,6 @@ public class SecurityGroupResource extends Ec2TaggableResource<SecurityGroup> im
         setOwnerId(group.ownerId());
         setDescription(group.description());
 
-        getEgress().clear();
-        for (IpPermission permission : group.ipPermissionsEgress()) {
-            for (IpRange ipRange : permission.ipRanges()) {
-                if (isKeepDefaultEgressRules() && permission.ipProtocol().equals("-1") && ipRange.cidrIp().equals("0.0.0.0/0")) {
-                    continue;
-                }
-
-                SecurityGroupEgressRuleResource rule = newSubresource(SecurityGroupEgressRuleResource.class);
-                rule.copyPortProtocol(permission);
-                rule.setCidrBlock(ipRange.cidrIp());
-                rule.setDescription(ipRange.description());
-                getEgress().add(rule);
-            }
-
-            for (Ipv6Range ipRange : permission.ipv6Ranges()) {
-                if (isKeepDefaultEgressRules() && permission.ipProtocol().equals("-1") && ipRange.cidrIpv6().equals("::/0")) {
-                    continue;
-                }
-
-                SecurityGroupEgressRuleResource rule = newSubresource(SecurityGroupEgressRuleResource.class);
-                rule.copyPortProtocol(permission);
-                rule.setIpv6CidrBlock(ipRange.cidrIpv6());
-                rule.setDescription(ipRange.description());
-                getEgress().add(rule);
-            }
-
-            for (UserIdGroupPair groupPair : permission.userIdGroupPairs()) {
-                SecurityGroupEgressRuleResource rule = newSubresource(SecurityGroupEgressRuleResource.class);
-                rule.copyPortProtocol(permission);
-                rule.setSecurityGroup(findById(SecurityGroupResource.class, groupPair.groupId()));
-                rule.setDescription(groupPair.description());
-                getEgress().add(rule);
-            }
-        }
-
-        getIngress().clear();
-        for (IpPermission permission : group.ipPermissions()) {
-            for (IpRange ipRange : permission.ipRanges()) {
-                SecurityGroupIngressRuleResource rule = newSubresource(SecurityGroupIngressRuleResource.class);
-                rule.copyPortProtocol(permission);
-                rule.setCidrBlock(ipRange.cidrIp());
-                rule.setDescription(ipRange.description());
-                getIngress().add(rule);
-            }
-
-            for (Ipv6Range ipRange : permission.ipv6Ranges()) {
-                SecurityGroupIngressRuleResource rule = newSubresource(SecurityGroupIngressRuleResource.class);
-                rule.copyPortProtocol(permission);
-                rule.setIpv6CidrBlock(ipRange.cidrIpv6());
-                rule.setDescription(ipRange.description());
-                getIngress().add(rule);
-            }
-
-            for (UserIdGroupPair groupPair : permission.userIdGroupPairs()) {
-                SecurityGroupIngressRuleResource rule = newSubresource(SecurityGroupIngressRuleResource.class);
-                rule.copyPortProtocol(permission);
-                rule.setSecurityGroup(findById(SecurityGroupResource.class, groupPair.groupId()));
-                rule.setDescription(groupPair.description());
-                getIngress().add(rule);
-            }
-        }
-
         refreshTags();
     }
 
@@ -287,23 +165,10 @@ public class SecurityGroupResource extends Ec2TaggableResource<SecurityGroup> im
         );
 
         setId(response.groupId());
-
-        if (!isKeepDefaultEgressRules()) {
-            deleteDefaultEgressRule(client);
-        }
     }
 
     @Override
     protected void doUpdate(GyroUI ui, State state, AwsResource config, Set<String> changedProperties) {
-        SecurityGroupResource current = (SecurityGroupResource) config;
-
-        if (!current.isKeepDefaultEgressRules() && isKeepDefaultEgressRules()) {
-            throw new GyroException("Default rule removed. Cannot be undone.");
-        }
-
-        if (current.isKeepDefaultEgressRules() && !isKeepDefaultEgressRules()) {
-            deleteDefaultEgressRule(createClient(Ec2Client.class));
-        }
     }
 
     @Override
@@ -359,33 +224,5 @@ public class SecurityGroupResource extends Ec2TaggableResource<SecurityGroup> im
         }
 
         return securityGroup;
-    }
-
-    private void deleteDefaultEgressRule(Ec2Client client) {
-        SecurityGroup group = getSecurityGroup(client);
-
-        IpPermission permission = group.ipPermissionsEgress().stream()
-            .filter(o -> o.ipProtocol().equals("-1") && o.ipRanges().stream().anyMatch(oo -> oo.cidrIp().equals("0.0.0.0/0")))
-            .findFirst().orElse(null);
-
-        if (permission != null) {
-            SecurityGroupEgressRuleResource defaultRule = newSubresource(SecurityGroupEgressRuleResource.class);
-            defaultRule.copyPortProtocol(permission);
-            defaultRule.setDescription(permission.ipRanges().stream().filter(o -> o.cidrIp().equals("0.0.0.0/0")).findFirst().get().description());
-            defaultRule.setCidrBlock("0.0.0.0/0");
-            defaultRule.delete(client, getId());
-        }
-
-        IpPermission permissionIpv6 = group.ipPermissionsEgress().stream()
-            .filter(o -> o.ipProtocol().equals("-1") && o.ipv6Ranges().stream().anyMatch(oo -> oo.cidrIpv6().equals("::/0")))
-            .findFirst().orElse(null);
-
-        if (permissionIpv6 != null) {
-            SecurityGroupEgressRuleResource defaultRule = newSubresource(SecurityGroupEgressRuleResource.class);
-            defaultRule.copyPortProtocol(permissionIpv6);
-            defaultRule.setDescription(permissionIpv6.ipv6Ranges().stream().filter(o -> o.cidrIpv6().equals("::/0")).findFirst().get().description());
-            defaultRule.setIpv6CidrBlock("::/0");
-            defaultRule.delete(client, getId());
-        }
     }
 }
